# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 重点项目科研管理系统 API
  version: 1.0.0
  description: |
    当前 React 原型对应的后端契约。ID 在 JSON 中统一使用 string；时间为 RFC 3339，日期为 ISO 8601。
    归档材料无审批流程；成果和月季报采用初审、终审并保留提交版本快照。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /archive/national:
    get:
      tags: [Archives]
      operationId: listNationalArchiveDirectories
      parameters: [{ name: topicId, in: query, schema: { type: string } }, { name: unitId, in: query, schema: { type: string } }]
      responses: { '200': { description: 按课题和单位隔离的国家材料目录, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ArchiveDirectory' } } } } } }
  /archive/national/topics/{topicId}/units/{unitId}/folders:
    get:
      tags: [Archives]
      operationId: listNationalArchiveFolders
      parameters: [{ $ref: '#/components/parameters/TopicId' }, { $ref: '#/components/parameters/UnitId' }]
      responses: { '200': { description: 固定要求和本单位自定义文件夹, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ArchiveFolder' } } } } } }
    post:
      tags: [Archives]
      operationId: createNationalCustomFolder
      description: 有权访问该目录的管理角色或本单位账号可创建；required=true表示必存，false表示有则必存。
      parameters: [{ $ref: '#/components/parameters/TopicId' }, { $ref: '#/components/parameters/UnitId' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string, maxLength: 200 }, required: { type: boolean, default: true } } } } } }
      responses: { '201': { description: 已创建单位私有文件夹, content: { application/json: { schema: { $ref: '#/components/schemas/ArchiveFolder' } } } } }
  /archive/folders/{folderId}:
    delete:
      tags: [Archives]
      operationId: deleteCustomArchiveFolder
      description: 只允许删除有权管理的空自定义文件夹；系统模板不可删除，技术负责人和科研助理创建的文件夹只有创建者可删除。
      parameters: [{ $ref: '#/components/parameters/FolderId' }]
      responses: { '204': { description: 已删除空的自定义文件夹 }, '409': { $ref: '#/components/responses/Conflict' } }
  /archive/folders/{folderId}/files:
    get:
      tags: [Archives]
      operationId: listArchiveFolderFiles
      parameters: [{ $ref: '#/components/parameters/FolderId' }]
      responses: { '200': { description: 文件列表, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/FileObject' } } } } } }
    post:
      tags: [Archives]
      operationId: attachFileToArchiveFolder
      parameters: [{ $ref: '#/components/parameters/FolderId' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [fileId], properties: { fileId: { type: string } } } } } }
      responses: { '201': { description: 已关联并直接计入归档完成度, content: { application/json: { schema: { $ref: '#/components/schemas/FileObject' } } } } }
  /archive/folders/{folderId}/files/{fileId}:
    delete:
      tags: [Archives]
      operationId: removeArchiveFolderFile
      parameters: [{ $ref: '#/components/parameters/FolderId' }, { $ref: '#/components/parameters/FileId' }]
      responses: { '204': { description: 已解除关联；物理文件仍可被其他业务引用 } }
  /self-funded-projects:
    get:
      tags: [Archives]
      operationId: listSelfFundedProjects
      parameters: [{ name: topicId, in: query, schema: { type: string } }, { name: unitId, in: query, schema: { type: string } }]
      responses: { '200': { description: 内部单位可见范围内的项目, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/SelfFundedProject' } } } } } }
    post:
      tags: [Archives]
      operationId: createSelfFundedProject
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SelfFundedProjectRequest' } } } }
      responses: { '201': { description: 已创建并冻结材料模板, content: { application/json: { schema: { $ref: '#/components/schemas/SelfFundedProject' } } } } }
  /self-funded-projects/{projectId}:
    get:
      tags: [Archives]
      operationId: getSelfFundedProject
      parameters: [{ name: projectId, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: 项目与材料完成度, content: { application/json: { schema: { $ref: '#/components/schemas/SelfFundedProject' } } } } }
    put:
      tags: [Archives]
      operationId: updateSelfFundedProject
      parameters: [{ name: projectId, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SelfFundedProjectRequest' } } } }
      responses: { '200': { description: 已更新, content: { application/json: { schema: { $ref: '#/components/schemas/SelfFundedProject' } } } } }
  /self-funded-projects/{projectId}/folders:
    get:
      tags: [Archives]
      operationId: listSelfFundedProjectFolders
      parameters: [{ name: projectId, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: 项目创建时冻结的文件夹清单, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ArchiveFolder' } } } } } }
    post:
      tags: [Archives]
      operationId: createSelfFundedCustomFolder
      parameters: [{ name: projectId, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string, maxLength: 200 }, required: { type: boolean, default: true } } } } } }
      responses: { '201': { description: 已创建自定义材料文件夹, content: { application/json: { schema: { $ref: '#/components/schemas/ArchiveFolder' } } } } }
  /archive-progress:
    get:
      tags: [Archives]
      operationId: getArchiveProgress
      parameters: [{ name: topicId, in: query, schema: { type: string } }, { name: unitId, in: query, schema: { type: string } }, { name: ownerType, in: query, schema: { type: string, enum: [TOPIC_NATIONAL, SELF_FUNDED] } }]
      responses: { '200': { description: 科研助理与技术负责人可见的归档统计, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ArchiveProgress' } } } } } }
  /files/upload-tickets:
    post:
      tags: [Files]
      operationId: createUploadTicket
      description: 返回短期上传地址；当前本地实现写入 FILE_STORAGE_ROOT 指定目录，数据库只保存元数据。存储位置通过 FileStorage 抽象可替换为 MinIO/S3。
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UploadTicketRequest' } } } }
      responses: { '201': { description: 上传票据, content: { application/json: { schema: { $ref: '#/components/schemas/UploadTicket' } } } } }
  '/files/{fileId}:complete':
    post:
      tags: [Files]
      operationId: completeUpload
      parameters: [{ $ref: '#/components/parameters/FileId' }]
      responses: { '200': { description: 校验对象后登记完成, content: { application/json: { schema: { $ref: '#/components/schemas/FileObject' } } } } }
  /files/{fileId}/download-url:
    get:
      tags: [Files]
      operationId: getDownloadUrl
      parameters: [{ $ref: '#/components/parameters/FileId' }]
      responses: { '200': { description: 短期下载地址, content: { application/json: { schema: { $ref: '#/components/schemas/SignedUrl' } } } } }
  /files/{fileId}/preview-url:
    get:
      tags: [Files]
      operationId: getPreviewUrl
      parameters: [{ $ref: '#/components/parameters/FileId' }]
      responses: { '200': { description: 短期预览地址, content: { application/json: { schema: { $ref: '#/components/schemas/SignedUrl' } } } } }
  /files/{fileId}/content:
    put:
      tags: [Files]
      operationId: uploadFileContent
      description: 使用上传票据的 expires 和 signature，携带登录令牌与票据指定的 Content-Type 上传原始字节；大小及可选 SHA-256 由服务端校验。
      parameters:
        - { $ref: '#/components/parameters/FileId' }
        - { name: expires, in: query, required: true, schema: { type: integer, format: int64 } }
        - { name: signature, in: query, required: true, schema: { type: string } }
      requestBody: { required: true, content: { application/octet-stream: { schema: { type: string, format: binary } } } }
      responses: { '204': { description: 字节已写入配置的文件目录；还需调用 complete } }
    get:
      tags: [Files]
      operationId: readFileContent
      description: 使用短期下载或预览地址，并携带登录令牌；仍会检查当前用户的数据范围和文件状态。仅 PDF 与常见栅格图片允许内联预览，其余类型强制作为附件下载。
      parameters:
        - { $ref: '#/components/parameters/FileId' }
        - { name: expires, in: query, required: true, schema: { type: integer, format: int64 } }
        - { name: signature, in: query, required: true, schema: { type: string } }
        - { name: preview, in: query, schema: { type: boolean, default: false } }
      responses: { '200': { description: 文件内容, content: { application/octet-stream: { schema: { type: string, format: binary } } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    UnitId: { name: unitId, in: path, required: true, schema: { type: string } }
    TopicId: { name: topicId, in: path, required: true, schema: { type: string } }
...
    FolderId: { name: folderId, in: path, required: true, schema: { type: string } }
    FileId: { name: fileId, in: path, required: true, schema: { type: string } }
  responses:
...
    Conflict: { description: 唯一约束、状态或乐观锁冲突, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, instance, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, example: TOPIC_NOT_OPERATIONAL }
        traceId: { type: string }
        fieldErrors: { type: array, items: { type: object, properties: { field: { type: string }, message: { type: string } } } }
...
    ArchiveDirectory:
      type: object
      required: [topicId, unitId, folderCount, requiredCount, completedCount, completionRate]
      properties: { topicId: { type: string }, topicName: { type: string }, unitId: { type: string }, unitName: { type: string }, folderCount: { type: integer }, requiredCount: { type: integer }, completedCount: { type: integer }, completionRate: { type: number } }
    ArchiveFolder:
      type: object
      required: [id, name, ownerType, required, requiredQuantity, custom, fileCount, completed, canDelete]
      properties: { id: { type: string }, topicId: { type: [string, 'null'] }, unitId: { type: [string, 'null'] }, ownerType: { type: string, enum: [TOPIC_NATIONAL, SELF_FUNDED] }, ownerId: { type: string }, name: { type: string }, required: { type: boolean }, requiredQuantity: { type: integer }, custom: { type: boolean }, fileCount: { type: integer }, completed: { type: boolean }, canDelete: { type: boolean } }
    SelfFundedProjectRequest:
      type: object
      required: [topicId, code, name, projectType, principalName, startDate, endDate]
      properties: { topicId: { type: string }, code: { type: string }, name: { type: string }, projectType: { type: string, enum: [TECHNOLOGY, RENOVATION, INFRASTRUCTURE] }, principalName: { type: string }, startDate: { type: string, format: date }, endDate: { type: string, format: date }, budget: { type: number }, status: { type: string }, recordVersion: { type: integer } }
    SelfFundedProject:
      type: object
      description: 历史项目可能尚未补齐日期；新建和编辑请求均要求开始、结束日期。
      required: [id, topicId, ownerUnitId, code, name, projectType, principalName, status, recordVersion, templateSnapshotId, completionRate]
      properties:
        id: { type: string }
        topicId: { type: string }
        ownerUnitId: { type: string }
        code: { type: string }
        name: { type: string }
        projectType: { type: string, enum: [TECHNOLOGY, RENOVATION, INFRASTRUCTURE] }
        principalName: { type: string }
        startDate: { type: [string, 'null'], format: date }
        endDate: { type: [string, 'null'], format: date }
        budget: { type: [number, 'null'] }
        status: { type: string }
        recordVersion: { type: integer }
        templateSnapshotId: { type: string }
        completionRate: { type: number }
    ArchiveProgress:
      type: object
      required: [topicId, unitId, ownerType, requiredCount, completedCount, completionRate]
      properties: { topicId: { type: string }, unitId: { type: string }, ownerType: { type: string }, requiredCount: { type: integer }, completedCount: { type: integer }, completionRate: { type: number } }
    UploadTicketRequest:
      type: object
      required: [fileName, size, contentType, businessType]
      properties: { fileName: { type: string }, size: { type: integer, format: int64, minimum: 1 }, contentType: { type: string }, sha256: { type: string }, businessType: { type: string, enum: [ACHIEVEMENT, ARCHIVE] } }
    UploadTicket:
      type: object
      required: [fileId, uploadUrl, method, expiresAt]
      properties: { fileId: { type: string }, uploadUrl: { type: string, format: uri }, method: { type: string, enum: [PUT, POST] }, headers: { type: object, additionalProperties: { type: string } }, expiresAt: { type: string, format: date-time } }
    FileObject:
      type: object
      required: [id, originalName, size, contentType, status, uploaderId, createdAt]
      properties: { id: { type: string }, originalName: { type: string }, size: { type: integer, format: int64 }, contentType: { type: string }, sha256: { type: string }, status: { type: string, enum: [PENDING, READY, DELETED] }, uploaderId: { type: string }, createdAt: { type: string, format: date-time } }
    SignedUrl:
      type: object
      required: [url, expiresAt]
      properties: { url: { type: string, format: uri }, expiresAt: { type: string, format: date-time } }
...